Improved input validation
digits other than 0, 1, & 2 are no longer considered valid in
guesses. Also made the app compile by commenting out lines of
code that caused errors and making other necessary changes.

Also implemented a random first guess. I changed the word list
to be an std::vector instead of an unordered_set because I
struggled to figure out how to get a random member of an unordered
set. It seemed a little complicated and I wasn't sure of the
benefit of doing that.
https://stackoverflow.com/questions/12761315/random-element-from-unordered-set-in-o1